pyproject.toml
[project]
name = "topdata-package-release-builder"
version = "0.1.0"
description = "Build and package Shopware 6 plugins for release"
requires-python = ">=3.8"
dependencies = [
    "rich>=13.0.0",
    "click>=8.0.0",
    "pytz",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
sw-build = "topdata_package_release_builder.cli:main"

